+++ src/methods/$includes-any-seq/$includes-any-seq.d.ts
@@ -1,0 +1,14 @@
+import { $Promise } from '../../../generate/async.macro';
+
+import { $SourceIterable } from '../../types/$iterable';
+
+declare function $includesAnySeq(
+  seqs: Array<$SourceIterable<any>>,
+): (iterable: $SourceIterable<any>) => $Promise<boolean>;
+
+declare function $includesAnySeq(
+  seqs: Array<$SourceIterable<any>>,
+  iterable: $SourceIterable<any>,
+): $Promise<boolean>;
+
+export default $includesAnySeq;