+++ src/methods/$includes-any/includes-any.mjs
@@ -19,7 +19,7 @@ export default iterableCurry(includesAny, {
   reduces: true,
   validateArgs(args) {
     if (true && typeof args[1] === 'string') {
-      console.warn(`For string inputs use includesAnySubseq instead of includesAny`);
+      console.warn(`For string inputs use includesAnySeq instead of includesAny`);
     }
   },
 });