+++ src/methods/$includes-seq/$includes-seq.d.ts
@@ -1,0 +1,14 @@
+import { $Promise } from '../../../generate/async.macro';
+
+import { $SourceIterable } from '../../types/$iterable';
+
+declare function $includesSeq(
+  seq: $SourceIterable<any>,
+): (iterable: $SourceIterable<any>) => $Promise<boolean>;
+
+declare function $includesSeq(
+  seq: $SourceIterable<any>,
+  iterable: $SourceIterable<any>,
+): $Promise<boolean>;
+
+export default $includesSeq;