+++ src/methods/$includes/includes.mjs
@@ -17,7 +17,7 @@ export default iterableCurry(includes, {
   reduces: true,
   validateArgs(args) {
     if (true && typeof args[1] === 'string') {
-      console.warn(`For string inputs use includesSubseq instead of includes`);
+      console.warn(`For string inputs use includesSeq instead of includes`);
     }
   },
 });