+++ src/methods/$starts-with-any-seq/$starts-with-any-seq.d.ts
@@ -1,0 +1,14 @@
+import { $Promise } from '../../../generate/async.macro';
+
+import { $SourceIterable } from '../../types/$iterable';
+
+declare function $startsWithAnySeq(
+  valueSeqs: Array<$SourceIterable<any>>,
+): (iterable: $SourceIterable<any>) => $Promise<boolean>;
+
+declare function $startsWithAnySeq(
+  valueSeqs: Array<$SourceIterable<any>>,
+  iterable: $SourceIterable<any>,
+): $Promise<boolean>;
+
+export default $startsWithAnySeq;