+++ src/methods/$starts-with-any/$starts-with-any.js
@@ -14,7 +14,7 @@ export default $iterableCurry($startsWithAny, {
   reduces: true,
   validateArgs(args) {
     if ($isSync && typeof args[1] === 'string') {
-      console.warn('For string inputs use startsWithAnySubseq instead of startsWithAny');
+      console.warn('For string inputs use startsWithAnySeq instead of startsWithAny');
     }
   },
 });